if phosh, do things
Disable rotation.
Fix the orientation to the best case.
Initialize window fullscreen.
Set the scale to a commond denominator of 768, a common resolution.
Use isPi hacks.
On close, put things back the way they were found.

This was implemented using a class, ui, with the intention of making it
easily expandedable for different UI's and devices.